migrate to sdl3

+++ src/Bar.hpp
@@ -1,9 +1,5 @@
 #pragma once
 
-#include <SDL.h>
-#include <SDL_rect.h>
-#include <SDL_render.h>
-
 #include "GameObject.hpp"
 
 enum class BarState {
@@ -14,7 +10,7 @@ enum class BarState {
 
 class Bar : public GameObject {
  public:
-  Bar(int x, int y, int width, int height, int red, int green, int blue,
+  Bar(float x, float y, float width, float height, int red, int green, int blue,
       int alpha, bool fill, BarState state);
   void draw();
   void render(SDL_Renderer *renderer);
