+++ src/GameObject.cpp
@@ -1,7 +1,7 @@
 #include "GameObject.hpp"
 
-GameObject::GameObject(int x, int y, int width, int height, int red, int green,
-                       int blue, int alpha, bool fill)
+GameObject::GameObject(float x, float y, float width, float height, int red,
+                       int green, int blue, int alpha, bool fill)
     : m_x{x},
       m_y{y},
       m_width{width},
@@ -13,10 +13,10 @@ GameObject::GameObject(int x, int y, int width, int height, int red, int green,
       m_fill{fill} {}
 
 // Getters
-int GameObject::get_x() { return m_x; }
-int GameObject::get_y() { return m_y; }
-int GameObject::get_width() { return m_width; }
-int GameObject::get_height() { return m_height; }
+float GameObject::get_x() { return m_x; }
+float GameObject::get_y() { return m_y; }
+float GameObject::get_width() { return m_width; }
+float GameObject::get_height() { return m_height; }
 int GameObject::get_red() { return m_red; }
 int GameObject::get_green() { return m_green; }
 int GameObject::get_blue() { return m_blue; }
@@ -24,10 +24,10 @@ int GameObject::get_alpha() { return m_alpha; }
 int GameObject::get_fill() { return m_fill; }
 
 // Setters
-void GameObject::set_x(int x) { m_x = x; }
-void GameObject::set_y(int y) { m_y = y; }
-void GameObject::set_width(int width) { m_width = width; }
-void GameObject::set_height(int height) { m_height = height; }
+void GameObject::set_x(float x) { m_x = x; }
+void GameObject::set_y(float y) { m_y = y; }
+void GameObject::set_width(float width) { m_width = width; }
+void GameObject::set_height(float height) { m_height = height; }
 void GameObject::set_red(int red) { m_red = red; }
 void GameObject::set_green(int green) { m_green = green; }
 void GameObject::set_blue(int blue) { m_blue = blue; }
