+++ src/Maze.hpp
@@ -1,8 +1,6 @@
 #pragma once
 
-#include <SDL.h>
-#include <SDL_rect.h>
-#include <SDL_render.h>
+#include <SDL3/SDL.h>
 
 #include <queue>
 #include <stack>
@@ -29,7 +27,7 @@ class Maze : public GameObject {
   // returns the cell id of the point(x, y)
   // the point(x, y) needs to be inside the maze
   // otherwise return {-1, -1}
-  std::pair<int, int> get_cell_id(int x, int y);
+  std::pair<int, int> get_cell_id(float x, float y);
   std::pair<int, int> get_source_cell_id();
   std::pair<int, int> get_target_cell_id();
   CellState get_cell_state(std::pair<int, int> cell_id);
@@ -59,7 +57,7 @@ class Maze : public GameObject {
 
   void dijkstra_ss();
   bool is_valid(std::pair<int, int> cell_id);
-  bool is_inside(int x, int y);
+  bool is_inside(float x, float y);
   bool is_not_visited(std::pair<int, int> cell_id);
   bool is_visited(std::pair<int, int> cell_id);
   void construct_shortest_path();
